Fix gesture list jumping sometimes if double clicked. Removed gesture list default sort.
The list is cleaned and rebuilt when there's a change, for example, a gesture is activated. And when a sort column is specified, the list is sorted if necessary just before the list is drawn. If a selection by mouse position is done before the sort is done, the wrong item might be selected.
I can't believe I spent 5 hours on this.